Co-locate observability handlers
