+++ apps/api/next.config.ts
@@ -1,8 +1,9 @@
 import { env } from '@/env';
-import { config, withAnalyzer, withSentry } from '@repo/next-config';
+import { config, withAnalyzer } from '@repo/next-config';
+import { withLogtail, withSentry } from '@repo/observability/next-config';
 import type { NextConfig } from 'next';
 
-let nextConfig: NextConfig = { ...config };
+let nextConfig: NextConfig = withLogtail({ ...config });
 
 if (env.VERCEL) {
   nextConfig = withSentry(nextConfig);
