+++ apps/app/env.ts
@@ -5,6 +5,7 @@ import { keys as database } from '@repo/database/keys';
 import { keys as email } from '@repo/email/keys';
 import { keys as core } from '@repo/env';
 import { keys as flags } from '@repo/feature-flags/keys';
+import { keys as observability } from '@repo/observability/keys';
 import { keys as security } from '@repo/security/keys';
 import { keys as webhooks } from '@repo/webhooks/keys';
 import { createEnv } from '@t3-oss/env-nextjs';
@@ -18,6 +19,7 @@ export const env = createEnv({
     database(),
     email(),
     flags(),
+    observability(),
     security(),
     webhooks(),
   ],
