+++ apps/app/next.config.ts
@@ -1,9 +1,10 @@
 import { env } from '@/env';
 import { withToolbar } from '@repo/feature-flags/lib/toolbar';
-import { config, withAnalyzer, withSentry } from '@repo/next-config';
+import { config, withAnalyzer } from '@repo/next-config';
+import { withLogtail, withSentry } from '@repo/observability/next-config';
 import type { NextConfig } from 'next';
 
-let nextConfig: NextConfig = withToolbar({ ...config });
+let nextConfig: NextConfig = withToolbar(withLogtail({ ...config }));
 
 if (env.VERCEL) {
   nextConfig = withSentry(nextConfig);
