+++ apps/web/next.config.ts
@@ -1,10 +1,11 @@
 import { env } from '@/env';
 import { withCMS } from '@repo/cms/next-config';
 import { withToolbar } from '@repo/feature-flags/lib/toolbar';
-import { config, withAnalyzer, withSentry } from '@repo/next-config';
+import { config, withAnalyzer } from '@repo/next-config';
+import { withLogtail, withSentry } from '@repo/observability/next-config';
 import type { NextConfig } from 'next';
 
-let nextConfig: NextConfig = withToolbar({ ...config });
+let nextConfig: NextConfig = withToolbar(withLogtail({ ...config }));
 
 nextConfig.images?.remotePatterns?.push({
   protocol: 'https',
