+++ packages/next-config/package.json
@@ -15,7 +15,6 @@
     "@logtail/next": "^0.1.7",
     "@next/bundle-analyzer": "^15.1.0",
     "@prisma/nextjs-monorepo-workaround-plugin": "^6.0.1",
-    "@sentry/nextjs": "^8.43.0",
     "next-secure-headers": "^2.2.0"
   }
 }
